Major clean up of txzchk - output + modularity (#65)
* Major clean up of txzchk - output + modularity

Show all warnings _after_ parsing files. This is done by saving every
line to another linked list so that after the entire tarball contents
are displayed (via the readline() function) the checks are run on the
list of lines rather than each line as the line is read. There are still
some warnings that are shown before and after rounding the total file(s)
size to a 1024 multiple but I tried to move as many warnings together as
possible (some warnings have to be detected whilst parsing each line;
others can be done at the end and some fit better when parsing each
line).

Move duplicate code into separate functions.

I tried also to make warning messages a more consistent format but more
can be done in this department.

These changes have been tested under linux (CentOS) and macOS with the
tarballs provided by the judges on GitHub as well as some text files
(via the -T option) I created and with the exception of subdirectories
(which I hope to work on later today) everything _should_ be detected.

* txzchk: detect directory change

* txzchk: report when correct directory not found

* Add check for subdirectories to txzchk

See the comments about how this is done:

    /* Check the path to see if there are any subdirectories. The way this is
     * done is counting the number of '/' but done carefully: for example the
     * path test-1//prog.c would not count as two directories but just one.
     *
     * Another example: ..// would be counted as one directory but it still has
     * ../ so that would have been detected above.
     *
     * It does this by saving the previous character: if it was also a '/' then
     * it's not counted as another directory: the first one will be counted
     * however since there wasn't one before it.
     *
     * We don't count the first character of the path because a path like:
     * /test-3/ would be counted as two directories but it's actually only one.
     * Well it kind of is two but / is special and it would still trigger an
     * absolute path warning - at least from a text file (tar strips it off) -
     * because it starts with a '/'.
     *
     * Note that the path /test-3 would trigger a warning that it's not in the
     * correct directory because it's an absolute directory. However since tar
     * strips the initial '/'s this would probably not get flagged.
     *
     * Note also that if the tar output does not have a trailing '/' in a
     * directory entry itself it would not count as another directory. However
     * since we also check for more than one 'd' line in the output it would
     * trigger more than one directory in the tarball.
     */

* txzchk: fix a false positive with > 1 directory

Specifically '././' counted as two because the '.' isn't a '/'. I'm
aware of at least one more thing that might be a problem.

* Update comments for previous commit