txzchk: check updates and fix dot file warning
Previously if no '/' was found in the path there would be no test for
'../' since it could not be found. Now however the latter check is for
'..' just in case someone manages to construct a tarball that ends with
'..'. Thus the code has now been changed to test for both. Dot files are
checked in the check_file() function.

Fix dot file warning: it used to say 'files' even if only one
unacceptable dot file was included; now it says 'file' if one dot file
and 'files' if more than 1.