txzchk: bug fixes and man page updates (#66)
* txzchk: strtok() -> strtok_r() to make it safer

* Added self-deprecating humour to txzchk.1

* Rename file_size variable to size

To remove ambiguity I renamed the variable in check_empty_file() as
there's also the file_size() function.

* txzchk: check updates and fix dot file warning

Previously if no '/' was found in the path there would be no test for
'../' since it could not be found. Now however the latter check is for
'..' just in case someone manages to construct a tarball that ends with
'..'. Thus the code has now been changed to test for both. Dot files are
checked in the check_file() function.

Fix dot file warning: it used to say 'files' even if only one
unacceptable dot file was included; now it says 'file' if one dot file
and 'files' if more than 1.

* Update txzchk.1 and iocccsize.1

Add to FILES section limit_ioccc.h file.

* Update date in iocccsize.1 and txzchk.1

Forgot to do this in previous commit.

* Version update in txzchk