Improved the way we maintain patches with the original iocccsize
We no longer maintain a copy of the original iocccsize tool.
Instead we did a clone of:

	https://github.com/SirWumpus/iocccsize

into an orig.iocccsize sub-directory that is not part of this repo.

Created a orig.iocccsize.ptch sub-directory where the patchs for
iocccsize.c and iocccsize-test.ptch reside.  Updated those patches
and the comments above those patches.