Initial JSON validator checks added + Makefile fix
Add the initial jauthchk and jinfochk tools: for now they only have the
help message, parse options and verify that the file: exists, is a
regular file and is readable. Because the structs are the same I have
moved the json specific structs (info, author) to json.h. This means
mkiocccentry.c has been updated to #include "json.h". Because struct
info is now in a header file I've also renamed the info struct in
txzchk.c to be instead txz_info.

Fix Makefile for make depend to work again (and run make depend itself).