Fixes in new JSON tools
Yesterday I started working on a general tool that was to check both
json files but when I split it off into two files I forgot to remove the
options (i.e. the parse_author_json and parse_info_json variables).

Additionally I added to both structs an 'issues' variable: these are
specific to the respective tools and now the return status of the tools
depend on the number of issues found. This means that now the structs
are not unused variables (which they were in the previous commit).

Finally I added a call to fclose() for each file that I forgot to do
yesterday.

Later today I hope to read the issues in full so I can start the actual
parsing of the files but this is a good beginning.