ssize_t -> off_t for file sizes + typo fix
Function file_size() now returns an off_t (as the st_size variable in
struct stat is also off_t).

All variables that are assigned the value of calls to file_size() are
now of off_t (txzchk.c already has this but mkiocccentry.c did not).
Note that since in the mkiocccentry tool the rule_2a_size (in struct
info) was assigned the value of a call to file_size() I also changed
this in struct info _and_ struct iocccsize; if this is wrong I will
happily fix it.

Note that in limit_ioccc.h the #define RULE_2A_SIZE had a cast to
ssize_t and I changed it to off_t; however in iocccsize.c itself in
stand-alone mode neither of the RULE_* defines have a cast; they're
just:

     #define RULE_2A_SIZE           4096    /* IOCCC Rule 2a */
     #define RULE_2B_SIZE           2503    /* IOCCC Rule 2b */

I am not sure if this is a problem but it's inconsistent with what it is
in limit_ioccc.h. I will mention this in a comment on GitHub to see if
it should also be fixed.

iocccsize.c: typo fix.