Add check for jinfochk/jauthchk to mkiocccentry
In other words the -j and -J options are implemented and in the
sanity_chk() function we check that they: exist, are regular files and
are executable. What's not done is to call them or print out the message
prior to executing them. On the other hand the usage messages have been
updated.