mkiocccentry now runs jauthchk and jinfochk
This commit means that the only thing left to do is the actual parsing
and checking the validity of the json files.

The tools are executed right after writing the respective file so that
if .info.json fails to pass it'll abort before even attempting to write
the .author.json file. This is to keep each tool together with the
writing of the file (the file is closed prior to running the test!).

Added -q option so that when run from mkiocccentry it doesn't show the
welcome string etc.

The man page for mkiocccentry still has to be updated for this but I'll
do that next.