Initial JSON tool checkers + add my URLs in txzchk.c (#70)
* Add my URLs below my handle and name

Thank you again Landon for adding my name to txzchk.c! I'm honoured and
privileged and I'm happy I could help!

* Initial JSON validator checks added + Makefile fix

Add the initial jauthchk and jinfochk tools: for now they only have the
help message, parse options and verify that the file: exists, is a
regular file and is readable. Because the structs are the same I have
moved the json specific structs (info, author) to json.h. This means
mkiocccentry.c has been updated to #include "json.h". Because struct
info is now in a header file I've also renamed the info struct in
txzchk.c to be instead txz_info.

Fix Makefile for make depend to work again (and run make depend itself).

* Fixes in new JSON tools

Yesterday I started working on a general tool that was to check both
json files but when I split it off into two files I forgot to remove the
options (i.e. the parse_author_json and parse_info_json variables).

Additionally I added to both structs an 'issues' variable: these are
specific to the respective tools and now the return status of the tools
depend on the number of issues found. This means that now the structs
are not unused variables (which they were in the previous commit).

Finally I added a call to fclose() for each file that I forgot to do
yesterday.

Later today I hope to read the issues in full so I can start the actual
parsing of the files but this is a good beginning.

* ssize_t -> off_t for file sizes + typo fix

Function file_size() now returns an off_t (as the st_size variable in
struct stat is also off_t).

All variables that are assigned the value of calls to file_size() are
now of off_t (txzchk.c already has this but mkiocccentry.c did not).
Note that since in the mkiocccentry tool the rule_2a_size (in struct
info) was assigned the value of a call to file_size() I also changed
this in struct info _and_ struct iocccsize; if this is wrong I will
happily fix it.

Note that in limit_ioccc.h the #define RULE_2A_SIZE had a cast to
ssize_t and I changed it to off_t; however in iocccsize.c itself in
stand-alone mode neither of the RULE_* defines have a cast; they're
just:

     #define RULE_2A_SIZE           4096    /* IOCCC Rule 2a */
     #define RULE_2B_SIZE           2503    /* IOCCC Rule 2b */

I am not sure if this is a problem but it's inconsistent with what it is
in limit_ioccc.h. I will mention this in a comment on GitHub to see if
it should also be fixed.

iocccsize.c: typo fix.

* txzchk: Add debug output + typecasts fix

New function 'show_txz_info()': reports on all the fields in txz_info
struct (depending on verbosity level). This includes some new fields
added as well as moving some of the global (and local) variables into
the struct.

The off_t variables (size of tarball, total file size, rounded file size
of all files) now are (for printing) cast to long long.

Updated version.

* Add check for jinfochk/jauthchk to mkiocccentry

In other words the -j and -J options are implemented and in the
sanity_chk() function we check that they: exist, are regular files and
are executable. What's not done is to call them or print out the message
prior to executing them. On the other hand the usage messages have been
updated.

* mkiocccentry now runs jauthchk and jinfochk

This commit means that the only thing left to do is the actual parsing
and checking the validity of the json files.

The tools are executed right after writing the respective file so that
if .info.json fails to pass it'll abort before even attempting to write
the .author.json file. This is to keep each tool together with the
writing of the file (the file is closed prior to running the test!).

Added -q option so that when run from mkiocccentry it doesn't show the
welcome string etc.

The man page for mkiocccentry still has to be updated for this but I'll
do that next.

* Update mkiocccentry.1 for jinfochk/jauthchk

* Add missing jinfochk.o to Makefile + make depend