txzchk: two additional checks added
txzchk now checks for invalid chars in filenames (full path): that is it
now only allows fully POSIX portable chars (see mkiocccentry.c) as well
as '/' (since we check on the full path and not just the basename). I.e.
it checks on the field 'char *filename' in struct file. Note that as
soon as an invalid char is detected it breaks out of the loop (so a file
with two invalid chars only triggers one warning and issue).

txzchk also now checks for files called just '.'. These files are
counted as BOTH an invalidly named dot file AND a file called just '.'
(which is clearly abuse).

The show_txz_info() function has been updated for both of these changes.
As well instead of showing if there were any unacceptable dot files it
shows how many.