Add jinfochk.1 and jauthchk.1
Note that for some of the text it has words like 'should' because these
tools are not finished (when it comes to mkiocccentry it already is
implemented so this wording is not there).

I did notice that the man pages do not have a SEE ALSO section. This
should be fixed but I'll worry about this at another time.