jauthchk, jinfochk, txzchk return value update
These tools now exit with the number of issues found. For example if
txzchk finds no problems in the tarball it'll be equivalent to exit(0);
if it finds 5 problems it'll be exit(5). If the tools encounter internal
errors or a parsing error (due to invalid input for example) the end is
still not reached.

I've updated the man pages for these tools to say this. Additionally
(for jinfochk.1 and jauthchk.1) I've added to the FILES section json.c.