Update .gitignore
Because I have a bunch of entry.test-*txz files (for when I was testing
txzchk - and it's useful when I add another test for example) I have
added this to the .gitignore: entry.*.tx[zt]: because I have to update
the file as txt files as well. But this way these files aren't
accidentally added to the repo. I realise I could use .git/info/exclude
but I feel this is a good place to have it for the above reason: so
nobody accidentally commits an entry tarball.