jauthchk, jinfochk, txzchk return value update (#74)
* Readd jauthchk.1 and jinfochk.1

When fixing the previous commits I forgot to do a git add for the two
man pages so they were not actually committed.

* jauthchk, jinfochk, txzchk return value update

These tools now exit with the number of issues found. For example if
txzchk finds no problems in the tarball it'll be equivalent to exit(0);
if it finds 5 problems it'll be exit(5). If the tools encounter internal
errors or a parsing error (due to invalid input for example) the end is
still not reached.

I've updated the man pages for these tools to say this. Additionally
(for jinfochk.1 and jauthchk.1) I've added to the FILES section json.c.

* Update .gitignore

Because I have a bunch of entry.test-*txz files (for when I was testing
txzchk - and it's useful when I add another test for example) I have
added this to the .gitignore: entry.*.tx[zt]: because I have to update
the file as txt files as well. But this way these files aren't
accidentally added to the repo. I realise I could use .git/info/exclude
but I feel this is a good place to have it for the above reason: so
nobody accidentally commits an entry tarball.

* Typo fix in json.c

'concert' -> 'convert'. I believe that '...character is not a value hex
character' should read 'valid hex character' but I'm not sure if I'm
misreading this so I've not changed it.

* Fix some (not all) of the warnings in json.c

Initialise chars to '\0' at time of declaration. There still are some
other warnings that have to be resolved (or dealt with).