Fix answers file issue
Having the #defines in the mkiocccentry.h caused a problem with the tool
being unable to find the answers file tags. I'm making this commit now
to make sure it's put in but the next commit I will have the other
defines in the header files moved to the respective source file.