Updates (#78)
* Typo fixes in util.c and Makefile

* Tidy up jstrencode, jstrdecode and mkiocccentry

Tidy up jstrencode, jstrdecode and mkiocccentry: Add jstrencode.h,
jstrdecode.h and mkiocccentry.h and update the C files and Makefile
(plus make depend) for this.

I also moved the usage string and REQUIRED_ARGS in txzchk.c to txzchk.h.

Uses include guards to prevent duplicate symbols.

* Fix answers file issue

Having the #defines in the mkiocccentry.h caused a problem with the tool
being unable to find the answers file tags. I'm making this commit now
to make sure it's put in but the next commit I will have the other
defines in the header files moved to the respective source file.

* Move #defines to source files

Although the #define REQUIRED_ARGS appears to be fine in the header file
I've moved the #defines back to the source files because of the problem
solved in the previous commit.