jinfochk and jauthchk updates
Add strict mode for testing of first char being a '{' and last char
being a '}'. In the previous commit the mkiocccentry tool uses this -s
option to these tools.

Add header files for jauthchk and jinfochk. Updated Makefile and run
make depend.

Add to json.h/json.c two new functions: check_first_json_char() and
check_last_json_char(): these functions check if the first char is a '{'
and last char is a '}' but with a strict boolean indicating how tight
the rules should be.

As they stand right now the jauthchk and jinfochk do the following:
check that the file can be opened and read; if not it's an error. Else
it reads it all in via the read_all() function. Assuming that function
does not return NULL the string is checked for embedded NUL bytes; if
any are found it is an error. Else strdup()d the string (for later).
Next it checks for the first character being '{' and last character
being '}' (both as discussed above). After that it frees all the strings
and closes the file. The remainder of the tests will follow in future
commits.