Minor updates to jinfochk and jauthchk
fclose() the respective file after reading in all the data, rather than
at the end of the function.

Don't use strlen() on the data read to see if it's 0 length; use the
length variable instead (in jauthchk.c this test was left out entirely).

Slightly change error message for embedded NUL bytes.