jinfochk and jauthchk updates (#79)
* Make file scope vars/functions static

When adding the header files for mkiocccentry, txzchk and jstrdecode
(not jstrencode) I omitted the static from the variables and functions.
This commit adds them back.

* jinfochk and jauthchk updates

Add strict mode for testing of first char being a '{' and last char
being a '}'. In the previous commit the mkiocccentry tool uses this -s
option to these tools.

Add header files for jauthchk and jinfochk. Updated Makefile and run
make depend.

Add to json.h/json.c two new functions: check_first_json_char() and
check_last_json_char(): these functions check if the first char is a '{'
and last char is a '}' but with a strict boolean indicating how tight
the rules should be.

As they stand right now the jauthchk and jinfochk do the following:
check that the file can be opened and read; if not it's an error. Else
it reads it all in via the read_all() function. Assuming that function
does not return NULL the string is checked for embedded NUL bytes; if
any are found it is an error. Else strdup()d the string (for later).
Next it checks for the first character being '{' and last character
being '}' (both as discussed above). After that it frees all the strings
and closes the file. The remainder of the tests will follow in future
commits.

* Minor updates to jinfochk and jauthchk

fclose() the respective file after reading in all the data, rather than
at the end of the function.

Don't use strlen() on the data read to see if it's 0 length; use the
length variable instead (in jauthchk.c this test was left out entirely).

Slightly change error message for embedded NUL bytes.

* Bump jauthchk version