Many checks added to jinfochk and jauthchk
The tools now extract fields and their values except that arrays are not
dealt with yet and not every other field is handled (many are handled).
For the field name remove leading and trailing whitespace and ONE '"'
from the beginning and end (I'm not sure if this is okay but it makes
parsing easier). For the values the removal of '"'s is only is done for
values that are not meant to be integers. It just occurred to me that
maybe this has to be done for booleans too but that can come in a later
commit.

jinfochk: test if our boolean values are in fact either "true" or
"false" (i.e. it uses strcmp()). This is not done in jauthchk because
.author.json has no booleans.

jinfochk and jauthchk: check for all the common fields except that
the handling of IOCCC_contest_id is NOT done yet (it is found but it is
not handled).

jinfochk: now accepts new option -F which allows you to specify the path
to the fnamchk tool (which is not used yet but will be in time).

Add function json_filename() which returns either ".info.json" or
".author.json" (depending on type passed in). If invalid type is passed
in return the string (not a NULL pointer!) "null". See json.h #defines
INFO_JSON and AUTHOR_JSON.

There might be some other changes I have forgotten.