Added IOCCC winner handle to .author.json
Added IOCCC winner handle to to each author.
The IOCCC winner handle will be used as part of
the rebuild of the www.ioccc.org web site.

A valid IOCCC winner handle must be of the form:

    [a-zA-Z0-9][a-zA-Z0-9._+-]

and must be <= MAX_WINNER_HANDLE characters long,
or is must be null.