mkiocccentry and txzchk: usage() function clean up
The usage() function in these tools were passed in the paths to the
tools the respective program used but these paths are actually just the
default paths so don't pass in those paths and instead just use them
directly in the usage() function in each file. The tools that had this
problem were mkiocccentry and txzchk; the others I already had done this
(or never had the problem).