Add tarball field to .author.json
This commit also adds -F /path/to/fnamchk to jauthchk (to verify that
the tarball field's value in the .author.json file is properly named).
Note that this tool is not yet used in either of the json checker tools
(jinfochk and jauthchk).

Add comment about not free()ing infop->tarball_path.

Whilst I often used the word 'whilst', I have changed all the uses of
the word 'whilst' to the word 'while' in the tools I have worked on
because I think 'while' is more natural than 'whilst' to many people.
Whilst this goes against my nature having 'while' instead of 'whilst'
seems a worthwhile thing since it is more natural to most people. :)