Write number of authors to JSON files
This will be used to verify that the number of authors is consistent
with the size of authors array (in .author.json): if the number of
authors is 2 and the authors array has 3 authors it's an error; if the
number of authors is 3 and the authors are numbers 0, 1 and 4 (three
authors but wrong numbers) there is a problem; etc.

In both write_author() and write_info() (the latter of which did not
previously take an int author_count) it is now an error if author_count
<= 0 (it was a warning before but if it is <= 0 it's a sign that there's
a serious problem because the code verifies that the number of authors
is within the allowed range which is > 0 && <= MAX_AUTHORS). Also these
functions now check if author_count > MAX_AUTHORS and if so this is also
an error (because the code does not allow it).

What I realised with this commit and adding the tarball path to the
.author.json (as requested) as well as the winner_handle is that the
files under test_JSON will fail eventually since the checkers will
expect to find the number of authors. Perhaps for the .author.json it's
a requirement and for the .info.json it can just warn? Or maybe there's
a way to easily add these fields to all those many test files?