Fix bug where asking for IOCCC winner handle fails
To test the jauthchk (when I get to the arrays part) I have made the
mkiocccentry-test.sh write five authors to the entry. But this in turn
revealed a bug: "IOCCC winner handle of author #3 does not start with an
alpha numeric character." I added the printing of the actual string
(seems helpful to the user) and it most certainly did start with an
alphanumerical character: it was "author3-last3". This commit fixes this
bug as well as (as above) having the maximum number of authors in the
make test script.