Two bug fixes in jinfochk and jauthchk
Array infinite loop when array is first: although arrays are not handled
yet this commit should prevent the tools from entering an infinite loop
when they are in the file ahead of everything else; it only happened if
the array was first because when extracting the value if it (the array)
wasn't first the strtok_r() call would be passed in NULL as expected:
when the array is first it happily continued to call strtok_r(p, ":",
&savefield); which means that p would once again be the same value until
the program was terminated.

Remove trailing whitespace from the value. This fixes the problem where
a valid value of false (as a boolean) triggered an error since using
strcmp() on it did not exactly match "false" (or "true"). Only happens
at the end of the file i.e. when not with a ',' at the end (this because
of the second argument passed into strtok_r() being ",\0" which does not
have a newline - which I think was the problem).