jinfochk and jauthchk: add IOCCC_contest_id check
Additionally the function get_contest_id() was passed a struct info *
but it was not used so I have removed it. The new function (in util.c)
'valid_contest_id()' cannot be used in get_contest_id() because the
latter gives out information to the user; it's interactive whereas the
valid_contest_id() isn't. This commit makes it so that every check
currently in check_common_json_fields() is done but the function has
more fields to be added later on.