Added -Werror to catch compile warngins
For now, added -Werror to make it easier to notice,
stop and fix when there is a compile warning.

As per Enhancement: Prepare for a formal release #87
this -Werror needs to be removed before a formal release
of the mkiocccentry tool set.  Once the 1st formal release
of the mkiocccentry tool set has been made, the -Werror
compile option should NOT be a default.