Add -T to all tools except iocccsize
The -T option in txzchk is now -f for regular file (-t was already taken
for tar path but to make the option consistent with the other tools I
named the option -T also which means I had to change the option char). I
had the thought of making it -x but I wasn't sure of that or the meaning
it might entail.

The man pages have all been updated for this change.