Various JSON updates, typo fixes + make seqcexit
check_common_json_fields() now takes a struct info * and struct author *
and depending on the program name it decides which can be NULL and which
cannot be. To do this the program_basename was added to jinfochk and
jauthchk. Note that these structs in the function are not yet used; they
will be in time but it requires a bit more thought and time. The idea
though is so that most (not all) issues can be reported at the end
rather than in the middle of parsing followed by aborting.

Move free_info() and free_author_array() to json.c: this is because
jinfochk and jauthchk will be using these structs more in a future
commit. Note that currently jauthchk does NOT call this function because
we have to first parse it to get the number of authors etc. This will
also require some thought and time; free_info() is called in jinfochk
though right now there's nothing to actually free in it (but since it's
all NULL it's safe).

The struct info char *tarball (which until earlier today was called
tarball_path) no longer shares memory with that of main() in
mkiocccentry. This is because in a future commit the tarball will be
added to the jinfochk and jauthchk tools.

Add /*ooo*/ to some lines in dbg.c to keep exit codes the same
(especially important as make test expects dbg_test to return with exit
code 5).

make seqcexit.