Update CHANGES.md
Since I added the -T option to (most of) the other tools I have noted
this in the file.

Additionally since the jinfochk and jauthchk tools have not been
finished I removed reference to their man pages since it appears that
the list was for tools that had been finished.