Add rule 2a/2b assignments in jinfochk
This only parses the values; it doesn't do anything else. For example if
the size is 0 it doesn't make sure that the override of empty prog.c is
true.

For the rule 2a (off_t) I use long long; for rule 2b (size_t) I use
unsigned long long. However this might be a problem so I am noting this
here but at least it's in now.