Move common JSON fields to json_common struct
That is to say that the fields that are written to .info.json and
.author.json are now in a common struct which is in both struct info and
struct author. Please do note that in mkiocccentry.c these still share
memory (or rather the info struct is used): previously in the functions
that refer to author and info structs it referenced e.g. info.ioccc_id
or infop->ioccc_id and the memory was only allocated in the info struct.
Now it's just info.common.ioccc_id or infop->common.ioccc_id instead.
This change is primarily for jinfochk and jauthchk in a future commit.