Start use of struct json_common
The function check_common_json_fields() now takes a struct json_common *
instead of either a struct info * or struct author *; this new struct is
in both struct info and struct author so depending on the tool it will
be the author copy or info copy. This will be useful when parsing the
fields for reporting after the parsing is completed. The function only
checks for NULL so far and the jinfochk and jauthchk pass in the correct
pointer; the function does not yet use the struct: that will happen
later.