Fix mkiocccentry -> jinfochk
With version update from commit 7abe7642fe41 make test no longer worked
as mkiocccentry failed after calling jinfochk:

	FATAL[223]: check_common_json_fields: iocccsize_version "28.8 2022-02-20" != IOCCCSIZE_VERSION "28.9 2022-02-23"
	FATAL[172]: write_info: ./jinfochk -q -s -F ./fnamchk -- test_work/test-0/.info.json failed with exit code: 223
	./mkiocccentry-test.sh: FATAL: mkiocccentry non-zero exit code: 172
	make: *** [test] Error 172

This is because mkiocccentry.c obtained the version from
iocccsize_version which is from rule_count and although it's set to
IOCCCSIZE_VERSION it did not seem to be picked up (from version.h). I'm
not sure why this is but mkiocccentry.h now #includes version.h and
mkiocccentry.c now refers to IOCCCSIZE_VERSION directly. Perhaps the
'char const * const iocccsize_version = IOCCCSIZE_VERSION;' should be
fixed but it was better to get the utility working than to worry about
that. Maybe it's also time to look into git hooks so that we verify that
make test works before a commit (though I'm not sure if that's possible
to configure for everyone or if it's only through client side
configuration)?