jinfochk/jauthchk: info and author versions checks
Previously I had not realised that IOCCC_info_version was not in the
.author.json file and instead it is IOCCC_author_version. Probably this
mistake was made from yank/paste but neither field is common to the
files. Thus the check for the former is removed from json.c and added to
jinfochk.c and the latter has been added to jauthchk.c.