jinfochk and jauthchk: warn instead of error
With the exception of actual parsing errors warn instead of issue an
error. The common fields are now added to a linked list that is iterated
through after parsing the file (the same will probably be done for the
other fields in time). The uncommon fields also only warn though fnamchk
will issue an error if the tarball is not correctly named. At the end
the number of issues found is returned to main() and main() returns 1 if
there were issues and otherwise 0.

This commit also fixes a bug in jinfochk: it always returned 0 because
the info struct was memset() to 0. As such main() in both tools now have
an int issues which keeps track of the number of issues instead.

What this change ultimately does is allow for proceeding even if an
issue is found; the exit code is still the same but it allows for an
overview of all issues (and makes it a lot easier to test multiple
problems in development).