Fix compilation under CentOS
This was due to the 'abort' bool in the new shell command functions when
abort() is also a function. Renamed it to 'abort_on_error'. Also
strangely I got unused variables which actually are used in the
respective file so I have removed the static. This might actually be
resolved by #ifdef FOO_C but I didn't think of this at the time - I
wanted to get this to work quickly before the pull request is merged.