Complete all checks on common json fields
That is to say that every field that is supposed to be in both
.info.json and .author.json are now checked that they're actually found
in the file being tested. Additionally it is reported if they're seen
more than once. This is done via a table with a count and a max allowed
plus a found bool. In actuality this is an array of struct json_fields
called 'common_json_fields'. When checking the found_common_json_fields
list if it's been seen more than once (common fields and most uncommon
fields are only allowed to be there once per file) the issues count is
incremented. Once the list is iterated through we go through all the
expected common fields and if they're not found in the
found_common_json_fields that's also an issue. This is why I added the
found_ to the names of the lists.

This commit does not do this for the uncommon fields yet. This will come
in a future commit. It might be that it can be done before arrays are
parsed (just that the array elements are not checked yet) but I'm not
sure of that either.