Comment fixes in json.c
Originally I named the 'loc' variable 'index' but I changed it to 'loc'
due to the index(3) but forgot to update the comment
(find_json_field_in_table()).

Also to clarify the last commit: in jinfochk and jauthchk if -t is
specified we don't worry about the versions of iocccsize, mkiocccentry
etc. This is because the test_JSON directories and their files were
added at different versions.